requirement.txt
fastapi
uvicorn
opencv-python
numpy
python-dotenv
google-genai
pytesseract
Pillow
httpx
paddleocr
paddlepaddle